serverless_s3/requirements.txt
awslambdaric==2.0.7
llama-cpp-python
boto3